requirements.txt
Flask
dnspython
requests